Move data to outside dir
