+++ src/pages/events.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Event, events } from './events-data';
+import { Event, events } from '../data/events.ts';
 import type { HeadFC, PageProps } from 'gatsby';
 import Layout from '../../components/Layout';
 import Card from '../../components/Card';
